chore(monitoring): refactor service & worker
